requirements.txt
numpy
scipy
pyrtlsdr
sounddevice
